Moved getUrlParams to utils folder
